Refactor known peers map to make room for additional fields
This is in service of eventually keeping track of more information about peers,
such as their identity payload. It will also allow us to get rid of the Instant
associated with all of the `PeerState` values right now, in favour of a dedicated
`last_updated` field on `PeerInfo` (coming soon).