A bit of refactoring to introduce handle_new_peer_discovered
We've got three places where we insert newly-discovered peers into the known_peers map.
This change refactors things to be a bit clearer and consistent, and introduces a new
function `handle_new_peer_discovered` that will be responsible for notifying consumers
of the library that a new peer has been discovered.